+++ demos/compose/src/main/java/androidx/media3/demo/compose/buttons/MinimalControls.kt
@@ -31,7 +31,8 @@ import androidx.media3.common.Player
 /**
  * Minimal playback controls for a [Player].
  *
- * Includes buttons for seeking to a previous/next items or playing/pausing the playback.
+ * Includes buttons for seeking to a previous/next items, skipping back/forward a couple of seconds,
+ * or playing/pausing the playback.
  */
 @Composable
 internal fun MinimalControls(player: Player, modifier: Modifier = Modifier) {
@@ -41,9 +42,13 @@ internal fun MinimalControls(player: Player, modifier: Modifier = Modifier) {
   Row(modifier = modifier, verticalAlignment = Alignment.CenterVertically) {
     Spacer(Modifier.weight(1f))
     PreviousButton(player, modifierForIconButton)
-    Spacer(Modifier.weight(0.5f))
+    Spacer(Modifier.weight(0.3f))
+    SeekBackButton(player, modifierForIconButton)
+    Spacer(Modifier.weight(0.3f))
     PlayPauseButton(player, modifierForIconButton)
-    Spacer(Modifier.weight(0.5f))
+    Spacer(Modifier.weight(0.3f))
+    SeekForwardButton(player, modifierForIconButton)
+    Spacer(Modifier.weight(0.3f))
     NextButton(player, modifierForIconButton)
     Spacer(Modifier.weight(1f))
   }